moved all packet builder methods to packetBuilder
